This appears to be stable, but does not work in most cases, so it is disabled
by default. Compile with --enable-avm2 to enable it.

The default build will now omit AVM2 code, which should reduce code size
and compile time. There is no warning yet for running AVM2 code without
AVM2 support.

The perl-generated haxe version of the actionscript.all testsuite is disabled
as Gnash fails some tests (tests may be bogus).

+++ libcore/SWFStream.cpp
@@ -23,6 +23,7 @@
 #include "IOChannel.h"
 #include "swf.h"
 #include "Property.h"
+#include "action_buffer.h"
 
 #include <cstring>
 #include <climits>
@@ -302,7 +303,7 @@ float    SWFStream::read_long_float()
 }
 
 // Read a 64-bit double value
-long double SWFStream::read_d64()
+double SWFStream::read_d64()
 {
 #ifdef USE_TU_FILE_BYTESWAPPING 
     align();
